categorize tests by action type
