+++ internal/e2e-realtime-api/src/voiceCollect/withDialListeners.test.ts
@@ -4,9 +4,9 @@ import {
   createTestRunner,
   CALL_COLLECT_PROPS,
   CALL_PROPS,
-  TestHandler,
   makeSipDomainAppAddress,
-} from './utils'
+  TestHandler,
+} from '../utils'
 
 const handler: TestHandler = ({ domainApp }) => {
   if (!domainApp) {
@@ -48,8 +48,8 @@ const handler: TestHandler = ({ domainApp }) => {
             // Wait until the caller starts the collect
             await waitForCollectStart
 
-            // Send wrong digits 123 to the caller (callee expects 1234)
-            const sendDigits = await call.sendDigits('1w2w3#')
+            // Send digits 1234 to the caller
+            const sendDigits = await call.sendDigits('1w2w3w4w#')
             tap.equal(
               call.id,
               sendDigits.id,
@@ -76,48 +76,45 @@ const handler: TestHandler = ({ domainApp }) => {
           domain: domainApp.domain,
         }),
         timeout: 30,
+        listen: {
+          onCollectStarted: (collect) => {
+            tap.hasProps(collect, CALL_COLLECT_PROPS, 'Collect started')
+            tap.equal(collect.callId, call.id, 'Collect correct call id')
+          },
+          onCollectInputStarted: (collect) => {
+            tap.hasProps(collect, CALL_COLLECT_PROPS, 'Collect input started')
+          },
+          // onCollectUpdated runs three times since callee sends 4 digits (1234)
+          // 4th (final) digit emits onCollectEnded
+          onCollectUpdated: (collect) => {
+            tap.hasProps(collect, CALL_COLLECT_PROPS, 'Collect updated')
+          },
+          onCollectFailed: (collect) => {
+            tap.notOk(collect.id, 'Collect failed')
+          },
+          onCollectEnded: (collect) => {
+            tap.hasProps(collect, CALL_COLLECT_PROPS, 'Collect ended')
+          },
+        },
       })
       tap.ok(call.id, 'Outbound - Call resolved')
 
-      const unsubCall = await call.listen({
-        onCollectStarted: (collect) => {
-          tap.hasProps(collect, CALL_COLLECT_PROPS, 'Collect started')
-          tap.equal(collect.callId, call.id, 'Collect correct call id')
-        },
-        onCollectInputStarted: (collect) => {
-          tap.hasProps(collect, CALL_COLLECT_PROPS, 'Collect input started')
-        },
-        // onCollectUpdated runs three times since callee sends 4 digits (1234)
-        // 4th (final) digit emits onCollectEnded
-        onCollectUpdated: (collect) => {
-          tap.hasProps(collect, CALL_COLLECT_PROPS, 'Collect updated')
-        },
-        onCollectFailed: (collect) => {
-          tap.notOk(collect.id, 'Collect failed')
-        },
-        onCollectEnded: (collect) => {
-          tap.hasProps(collect, CALL_COLLECT_PROPS, 'Collect ended')
-        },
-      })
-
       // Caller starts a collect
-      const collect = await call
-        .collect({
-          initialTimeout: 4.0,
-          digits: {
-            max: 4,
-            digitTimeout: 10,
-            terminators: '#',
-          },
-          partialResults: true,
-          continuous: false,
-          sendStartOfInput: true,
-          startInputTimers: false,
-        })
-        .onStarted()
+      const collect = call.collect({
+        initialTimeout: 4.0,
+        digits: {
+          max: 4,
+          digitTimeout: 10,
+          terminators: '#',
+        },
+        partialResults: true,
+        continuous: false,
+        sendStartOfInput: true,
+        startInputTimers: false,
+      })
       tap.equal(
         call.id,
-        collect.callId,
+        await collect.callId,
         'Outbound - Collect returns the same call instance'
       )
 
@@ -128,7 +125,7 @@ const handler: TestHandler = ({ domainApp }) => {
 
       // Compare what caller has received
       const recDigits = await collect.ended()
-      tap.not(recDigits.digits, '1234', 'Outbound - Received the same digit')
+      tap.equal(recDigits.digits, '1234', 'Outbound - Received the same digit')
 
       // Resolve the collect end promise
       waitForCollectEndResolve!()
@@ -151,13 +148,11 @@ const handler: TestHandler = ({ domainApp }) => {
 
       await unsubVoice()
 
-      await unsubCall()
-
       await client.disconnect()
 
       resolve(0)
     } catch (error) {
-      console.error('VoiceCollectCallListeners error', error)
+      console.error('VoiceCollectDialListeners error', error)
       reject(4)
     }
   })
@@ -165,7 +160,7 @@ const handler: TestHandler = ({ domainApp }) => {
 
 async function main() {
   const runner = createTestRunner({
-    name: 'Voice Collect with Call Listeners E2E',
+    name: 'Voice Collect with Dial Listeners E2E',
     testHandler: handler,
     executionTime: 60_000,
     useDomainApp: true,
