+++ internal/e2e-realtime-api/src/voicePrompt/withAllListeners.test.ts
@@ -6,7 +6,7 @@ import {
   CALL_PROMPT_PROPS,
   TestHandler,
   makeSipDomainAppAddress,
-} from './utils'
+} from '../utils'
 
 const handler: TestHandler = ({ domainApp }) => {
   if (!domainApp) {
