+++ internal/e2e-realtime-api/src/voicePrompt/withCallListeners.test.ts
@@ -7,7 +7,7 @@ import {
   CALL_PROMPT_PROPS,
   TestHandler,
   makeSipDomainAppAddress,
-} from './utils'
+} from '../utils'
 
 const handler: TestHandler = ({ domainApp }) => {
   if (!domainApp) {
